Fix the issues i got feedback on
