--- src/components/atoms/AchievementComponent/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AchievementComponent";
