+++ src/components/atoms/SpecialServiceCard/SpecialServiceCard.module.scss
@@ -3,17 +3,18 @@
 
 .service-card {
   @include center;
+
   flex-direction: column;
-  
-  gap: 15px;
   font-family: $playfair;
+  gap: 15px;
 
   &__title {
     color: $special-service-heading-color;
-    @include center;
     font-size: $fs-xxl;
-   @include margin-bottom($spacing);
     font-weight: $fw-bold;
+
+    @include center;
+   @include margin-bottom($spacing);
   }
 
   &__description {
